+++ test/tracing/opentracing/opentracingApi_test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var instanaOpentracing = require('../../../src/tracing/opentracing');
-var apiCompatibilityChecks = require('opentracing/lib/test/api_compatibility').apiCompatibilityChecks;
+var apiCompatibilityChecks = require('opentracing/lib/test/api_compatibility').default;
 
 describe('tracing/opentracing/opentracingApi', function() {
   apiCompatibilityChecks(instanaOpentracing.createTracer, {