+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## unreleased
- - Add tracing support for mysql2/promises.
+## Unreleased
+ - ioRedis instrumentation can lead to errors when async correlation doesn't work as expected.
+ - Add tracing support for mysql2/promise.
  
 ## 1.37.0
  - Enable tracing in Node.js `^9.1.0`.