+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
  - ioRedis instrumentation can lead to errors when async correlation doesn't work as expected.
  - Add tracing support for mysql2/promise.
- 
+ - Switch to `@risingstack/v8-profiler` due to security issues in transitive dependencies of `v8-profiler`.
+
 ## 1.37.0
  - Enable tracing in Node.js `^9.1.0`.